fix warning owing to DPL_APPEND_STR

+++ libdroplet/include/droplet/s3/reqbuilder.h
@@ -42,7 +42,7 @@ typedef enum
 /* PROTO reqbuilder.c */
 /* src/reqbuilder.c */
 dpl_status_t dpl_s3_req_build(const dpl_req_t *req, dpl_s3_req_mask_t req_mask, dpl_dict_t **headersp);
-dpl_status_t dpl_s3_req_gen_url(const dpl_req_t *req, dpl_dict_t *headers, char *buf, int len, unsigned int *lenp);
+dpl_status_t dpl_s3_req_gen_url(const dpl_req_t *req, dpl_dict_t *headers, char *buf, size_t len, unsigned int *lenp);
 dpl_status_t dpl_s3_add_authorization_to_headers(const dpl_req_t *, dpl_dict_t *,
                                                  const dpl_dict_t *, struct tm *);
